+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - ReleaseDate
 
+## [1.4.2] - 2024-02-17
+- Use more realistic sleep distribution in Terraform module
+- Make the number of processed resources finite in Terraform module - before it never actually quit
+
+## [1.4.1] - 2024-02-17
+- Fix hang in Terraform module on WASM [#624](https://github.com/svenstaro/genact/issues/624)
+
 ## [1.4.0] - 2024-02-16
 - Add Terraform module [#622](https://github.com/svenstaro/genact/pull/622) (thanks @taskinen)
 
@@ -70,7 +77,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Rewrite web version using wasm-bindgen
 
 <!-- next-url -->
-[Unreleased]: https://github.com/svenstaro/genact/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/svenstaro/genact/compare/v1.4.2...HEAD
+[1.4.2]: https://github.com/svenstaro/genact/compare/v1.4.1...v1.4.2
+[1.4.1]: https://github.com/svenstaro/genact/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/svenstaro/genact/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/svenstaro/genact/compare/v1.2.2...v1.3.0
 [1.2.2]: https://github.com/svenstaro/genact/compare/v1.2.1...v1.2.2